+++ lib/config/feature_flag.dart
@@ -5,4 +5,6 @@ class FeatureFlag {
   static const bool enableAutoSave = true;
   // enable Sync Layer to firebase
   static const bool enableSyncLayer = true;
+  // enable inapp update
+  static const bool enableInAppUpdate = true;
 }
